+++ NEWS.md
@@ -10,6 +10,7 @@
 * Added embed argument to `obs2gfm()` function
 * Created `sysdata.rda` to store data accessible with `::`
 * Added na values to color themes
+* Illumina codes
 
 # jakR2 0.3